// src/definitions/Response.d.ts
// Note: This file was automatically generated by https://transform.tools/json-to-typescript

export interface Response {
  facet_counts: unknown[];
  found: number;
  hits: Hit[];
  out_of: number;
  page: number;
  request_params: RequestParams;
  search_cutoff: boolean;
  search_time_ms: number;
}

export interface Hit {
  document: Document;
  highlight: Highlight;
  highlights: Highlights[];
  text_match: number;
  text_match_info: TextMatchInfo;
}

export interface Document {
  artist: string;
  artist_slug?: string;
  sfx_id?: number;
  featured_tags?: FeaturedTag[];
  id: string;
  image: string;
  is_explicit: number;
  keywords: string[];
  musicvine_track_id: number;
  name: string;
  premium: boolean;
  slug: string;
  styles: string[];
  tags: string[];
  track_tags: TrackTag[];
  version_length: number;
  version_precise_length: number;
  version_preview_uri: string;
  waveform: string;
}

export interface FeaturedTag {
  id: string;
  is_qualified: number;
  name: string;
  slug: string;
}

export interface TrackTag {
  id: string;
  is_qualified: number;
  name: string;
  slug: string;
}

export interface Highlight {
  name: Name;
}

export interface Name {
  matched_tokens: string[];
  snippet: string;
}

export interface Highlights {
  field: string;
  matched_tokens: string[];
  snippet: string;
}

export interface TextMatchInfo {
  best_field_score: string;
  best_field_weight: number;
  fields_matched: number;
  score: string;
  tokens_matched: number;
}

export interface RequestParams {
  collection_name: string;
  per_page: number;
  q: string;
}